+++ magento1.x/src/app/code/community/Loginradius/Sociallogin/Helper/LoginRadiusSDK/CustomerRegistration/AccountAPI.php
@@ -44,8 +44,7 @@ class AccountAPI
     public function accountLink($uid, $id, $provider)
     {
         $data = array('accountid' => $uid, 'provider' => $provider, 'providerid' => $id);
-        
-        return $this->apiClientHandler("link", array(), array('method' => 'post', 'post_data' => $data));
+        return $this->apiClientHandler("link", array(), array('method' => 'POST', 'post_data' => $data));
     }
 
     /**
@@ -60,7 +59,7 @@ class AccountAPI
     public function accountUnlink($uid, $id, $provider)
     {
         $data = array('accountid' => $uid, 'provider' => $provider, 'providerid' => $id);
-        return $this->apiClientHandler("unlink", array(), array('method' => 'post', 'post_data' => $data));
+        return $this->apiClientHandler("unlink", array(), array('method' => 'POST', 'post_data' => $data));
     }
 
     /**
@@ -76,7 +75,7 @@ class AccountAPI
      */
     public function createUserRegistrationProfile($data)
     {
-        return $this->apiClientHandler("profile", array(), array('method' => 'post', 'post_data' => $data, 'content_type' => 'json'));
+        return $this->apiClientHandler("profile", array(), array('method' => 'POST', 'post_data' => $data, 'content_type' => 'json'));
     }
 
     /**
@@ -88,7 +87,6 @@ class AccountAPI
      */
     public function getAccounts($uid)
     {
-        
         return $this->apiClientHandler("", array('accountid' => $uid));
     }
 
@@ -107,7 +105,7 @@ class AccountAPI
      */
     public function userAdditionalEmail($uid, $action, $data)
     {
-        return $this->apiClientHandler("email", array('accountid' => $uid, 'action' => $action), array('method' => 'post', 'post_data' => $data, 'content_type' => 'json'));
+        return $this->apiClientHandler("email", array('accountid' => $uid, 'action' => $action), array('method' => 'POST', 'post_data' => $data, 'content_type' => 'json'));
     }
 
     /**
@@ -146,7 +144,7 @@ class AccountAPI
     public function changeAccountPassword($uid, $old_password, $new_password)
     {
         $data = array('oldpassword' => $old_password, 'newpassword' => $new_password);
-        return $this->apiClientHandler("password", array('accountid' => $uid), array('method' => 'post', 'post_data' => $data));
+        return $this->apiClientHandler("password", array('accountid' => $uid), array('method' => 'POST', 'post_data' => $data));
     }
 
     /**
@@ -172,7 +170,7 @@ class AccountAPI
     public function setPassword($uid, $password)
     {
         $parameter = array('accountid' => $uid, 'action' => 'set');
-        return $this->apiClientHandler("password", $parameter, array('method' => 'post', 'post_data' => array('password' => $password)));
+        return $this->apiClientHandler("password", $parameter, array('method' => 'POST', 'post_data' => array('password' => $password)));
     }
 
     /**
@@ -187,7 +185,7 @@ class AccountAPI
     public function changeUsername($uid, $username, $new_username)
     {
         $data = array('currentusername' => $username, 'newusername' => $new_username);
-        return $this->apiClientHandler("changeusername", array('accountid' => $uid), array('method' => 'post', 'post_data' => $data));
+        return $this->apiClientHandler("changeusername", array('accountid' => $uid), array('method' => 'POST', 'post_data' => $data));
     }
 
     /**
@@ -210,7 +208,7 @@ class AccountAPI
      */
     public function setUsername($uid, $newusername)
     {
-        return $this->apiClientHandler("setusername", array('accountId' => $uid), array('method' => 'post', 'post_data' => array('newusername' => $newusername)));
+        return $this->apiClientHandler("setusername", array('accountId' => $uid), array('method' => 'POST', 'post_data' => array('newusername' => $newusername)));
     }
 
     /**
@@ -237,7 +235,7 @@ class AccountAPI
      */
     public function setStatus($uid, $action = true)
     {
-        return $this->apiClientHandler("status", array('accountid' => $uid), array('method' => 'post', 'post_data' => array('isblock' => $action)));
+        return $this->apiClientHandler("status", array('accountid' => $uid), array('method' => 'POST', 'post_data' => array('isblock' => $action)));
     }
 
     /**