+++ magento1.x/src/app/code/community/Loginradius/Sociallogin/Helper/SDKClient.php
@@ -51,7 +51,7 @@ class Loginradius_Sociallogin_Helper_SDKClient implements \LoginRadiusSDK\Client
             }
             $logData = array(
                 'url' => $request_url,
-                'method' => strtoupper(isset($options['method']) ? strtolower($options['method']) : 'get'),
+                'method' => strtoupper(isset($options['method']) ? strtoupper($options['method']) : 'GET'),
                 'data' => array(
                     'GET' => $query_array,
                     'POST' => (isset($options['post_data']) ? $options['post_data'] : array())
@@ -79,7 +79,7 @@ class Loginradius_Sociallogin_Helper_SDKClient implements \LoginRadiusSDK\Client
      */
     private function curlApiMethod($request_url, $options = array()) {
         $ssl_verify = isset($options['ssl_verify']) ? $options['ssl_verify'] : false;
-        $method = isset($options['method']) ? strtolower($options['method']) : 'get';
+        $method = isset($options['method']) ? strtoupper($options['method']) : 'GET';
         $data = isset($options['post_data']) ? $options['post_data'] : array();
         $content_type = isset($options['content_type']) ? trim($options['content_type']) : 'x-www-form-urlencoded';
         $curl_handle = curl_init();
@@ -89,7 +89,7 @@ class Loginradius_Sociallogin_Helper_SDKClient implements \LoginRadiusSDK\Client
         curl_setopt($curl_handle, CURLOPT_SSL_VERIFYPEER, $ssl_verify);
         if (!empty($data) || $data === true) {
             curl_setopt($curl_handle, CURLOPT_HTTPHEADER, array('Content-type: application/' . $content_type));
-            if ($method == 'post') {
+            if ($method == 'POST') {
                 curl_setopt($curl_handle, CURLOPT_POST, 1);
                 curl_setopt($curl_handle, CURLOPT_POSTFIELDS, (($content_type == 'json') ? json_encode($data) : LoginRadiusSDK\LoginRadius::queryBuild($data)));
             }
@@ -100,10 +100,6 @@ class Loginradius_Sociallogin_Helper_SDKClient implements \LoginRadiusSDK\Client
             curl_setopt($curl_handle, CURLOPT_RETURNTRANSFER, true);
         } else {
            
-            //curl_setopt($curl_handle, CURLOPT_HEADER, 1);
-            //$effectiveUrl = curl_getinfo($curl_handle, CURLINFO_EFFECTIVE_URL);
-            //curl_close($curl_handle);
-            //$curl_handle = curl_init();
             $url = str_replace('?', '/?', $request_url);
             
             curl_setopt($curl_handle, CURLOPT_URL, $url);
@@ -124,13 +120,13 @@ class Loginradius_Sociallogin_Helper_SDKClient implements \LoginRadiusSDK\Client
      */
     private function fsockopenApiMethod($request_url, $options = array()) {
         $ssl_verify = isset($options['ssl_verify']) ? $options['ssl_verify'] : false;
-        $method = isset($options['method']) ? strtolower($options['method']) : 'get';
+        $method = isset($options['method']) ? strtoupper($options['method']) : 'GET';
         $data = isset($options['post_data']) ? $options['post_data'] : array();
         $content_type = isset($options['content_type']) ? $options['content_type'] : 'form_params';
         if (!empty($data)) {
             $options = array('http' =>
                 array(
-                    'method' => strtoupper($method),
+                    'method' => $method,
                     'timeout' => 50,
                     'header' => 'Content-type :application/' . $content_type,
                     'content' => (($content_type == 'json') ? json_encode($data) : LoginRadiusSDK\LoginRadius::queryBuild($data))