+++ magento2.x/src/LoginRadius/Activation/view/adminhtml/web/activation/js/activation.js
@@ -1,0 +1,6 @@
+require(['jquery', 'jquery/ui'], function ($) {
+    $(document).ready(function () {
+        $('#lractivation_activationhelp,#lractivation_activationhelp-head').hide();
+    });
+    
+});