+++ magento2.x/src/LoginRadius/Apilog/Block/ApiLog.php
@@ -1,0 +1,11 @@
+<?php
+namespace LoginRadius\Apilog\Block;
+
+class Apilog extends \Magento\Framework\View\Element\Template
+{
+	public function _prepareLayout()
+	{
+		return parent::_prepareLayout();
+	}
+	
+}